updates
merged:

  gentoo-staging: fc74d3206fa20caa19b7703aa051ff6de95d5588
  flora: 0a1001b242b677fc7fae82003c519711485b70b9
  faustoo: fa7aa3fbcc892ef257664844cb32272583e3df02
  fusion809: 9daec531a164d89ba52d906900d871fddfb92e63
  kit-fixups: c3a5a7edc95a7b90e6e11d2900c97a20487db656

+++ x11-themes/redhat-artwork/redhat-artwork-5.0.8-r4.ebuild
@@ -1,5 +1,6 @@
-# Copyright 1999-2017 Gentoo Foundation
+# Copyright 1999-2014 Gentoo Foundation
 # Distributed under the terms of the GNU General Public License v2
+# $Id$
 
 EAPI=5
 inherit autotools eutils rpm
@@ -12,7 +13,7 @@ LICENSE="GPL-2"
 
 SLOT="0"
 KEYWORDS="alpha amd64 ~hppa ia64 ppc sparc x86"
-IUSE="audacious cursors gdm icons nautilus"
+IUSE="audacious cursors gdm icons kdm nautilus"
 
 RDEPEND="x11-libs/gtk+:2"
 DEPEND="
@@ -133,7 +134,7 @@ src_install () {
 	# Some extra features - allows redhat-artwork to be very light:
 	###
 	if ! use gdm; then rm -r "${D}"/usr/share/gdm || die; fi
-	rm -r "${D}"/usr/share/apps/kdm || die
+	if ! use kdm; then rm -r "${D}"/usr/share/apps/kdm || die; fi
 	if ! use cursors; then rm -r "${D}"/usr/share/cursors || die; fi
 	if ! use icons; then
 		rm -r "${D}"/usr/share/icons || die
